techflex-cms/requirements.txt
django==5.0
djangorestframework==3.15
drf-spectacular==0.27.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg[binary]==3.1.16
django-storages[boto3]==1.14.2
gunicorn==21.2.0
django-ckeditor==6.7.0
